REMOTE_MCP_SERVER/requirements.txt
# MCP 
fastmcp
aiohttp
python-dotenv

# Oracle DB 
oracledb